logging: Respect users pytest logging config
Reworks the way that step and console logging is done with the intention
of respecting the users pytest logging configuration, namely the
--log-(cli-|file-)(format|level) command line and corresponding
pytest.ini options.

This works by having the StepFormatter class take a "parent" argument
which is pytests original formatter, and having it pass logs to it
(after some munging) instead of doing formatting itself. For this to
work as intended, and also so step and console logging works *without*
the StepFormatter class, the formatting of log messages is now done when
the messages are logged, instead of in the StepFormatted. This makes the
"$(message)s" specifier in the formater handler work as expected.

The StepFormatter adds (optional) indentation behavior on top of this to make
logs prettier. Indentation is configured by passing extra properties
when constructing the log messages. These properties are handled by the
StepFromatter and indicate how much the current message should be
indented, and how much any subsequent messages should be indented if
they do not otherwise specify an indentation level.

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>